refactor: move AI tool tests to dedicated file and refactor markdown toolkit wrapper
